requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary==2.8.6
django-cors-headers